started working on paw serializer
