+++ src/serializers/paw/Serializer.js
@@ -1,1285 +1,471 @@
-import Context, {
-    Parameter
-} from '../../models/Core'
+import { List } from 'immutable'
+import { DynamicValue, DynamicString, RecordParameter } from '../../mocks/PawShims'
 
-import LateResolutionReference from '../../models/references/LateResolution'
-import JSONSchemaReference from '../../models/references/JSONSchema'
-import Reference from '../../models/references/Reference'
+import Store from '../../models/Store'
+import Auth from '../../models/Auth'
+import Reference from '../../models/Reference'
 
-import {
-    ApiKeyAuth
-} from '../../models/Auth'
+import { currify } from '../../utils/fp-utils'
 
-import Item from '../../models/Item'
-import ContextResolver from '../../resolvers/ContextResolver'
-import PawEnvironment from '../../models/environments/PawEnvironment'
-import ResolverOptions from '../../models/options/ResolverOptions'
+const __inputs__ = []
 
-import {
-    DynamicValue,
-    DynamicString,
-    InputField
-} from '../../mocks/PawShims'
+const methods = {}
 
-export default class PawSerializer {
-    static fileExtensions = [];
+export class PawSerializer {
+  static fileExtensions = [];
 
-    static inputs = [
-        new InputField(
-            'jsfInProtocol',
-            'Replace schemas by randomizers in protocol params',
-            'Checkbox',
-            { defaultValue: false, persisted: true }
-        ),
-        new InputField(
-            'jsfInHost',
-            'Replace schemas by randomizers in host params',
-            'Checkbox',
-            { defaultValue: false, persisted: true }
-        ),
-        new InputField(
-            'jsfInPath',
-            'Replace schemas by randomizers in path params',
-            'Checkbox',
-            { defaultValue: true, persisted: true }
-        ),
-        new InputField(
-            'jsfInQuery',
-            'Replace schemas by randomizers in query params',
-            'Checkbox',
-            { defaultValue: true, persisted: true }
-        ),
-        new InputField(
-            'jsfInHeaders',
-            'Replace schemas by randomizers in headers',
-            'Checkbox',
-            { defaultValue: true, persisted: true }
-        ),
-        new InputField(
-            'jsfInBody',
-            'Replace schemas by randomizers in body',
-            'Checkbox',
-            { defaultValue: true, persisted: true }
-        ),
-        new InputField(
-            'jsfInEnv',
-            'Replace schemas by randomizers in Environments',
-            'Checkbox',
-            { defaultValue: true, persisted: true }
-        )
-    ];
+  static inputs = __inputs__
 
-    constructor(context) {
-        this.ENVIRONMENT_DOMAIN_NAME = 'Imported Environments'
-        this.currentEnvironmentDomainName = null
-        this.context = context || null
-    }
-
-    _parserErrorHandler(e) {
-        /* eslint-disable no-console */
-        console.error(
-            '@parser failed with error',
-            e,
-            JSON.stringify(e),
-            e.stack
-        )
-        /* eslint-enable no-console */
-        throw e
-    }
-
-    _parserFailureHandler(e) {
-        /* eslint-disable no-console */
-        console.error(
-            '@parser caught error',
-            e,
-            JSON.stringify(e),
-            e.stack
-        )
-        /* eslint-enable no-console */
-        throw e
-    }
-
-    _serializerFailureHandler(e) {
-        /* eslint-disable no-console */
-        console.error(
-            '@serializer failed with error',
-            e,
-            JSON.stringify(e),
-            e.stack
-        )
-        throw e
-        /* eslint-enable no-console */
-    }
-
-    _resolverErrorHandler(error) {
-        /* eslint-disable no-console */
-        console.error(
-            '@resolver failed with error',
-            error,
-            JSON.stringify(error),
-            error.stack
-        )
-        throw error
-        /* eslint-enable no-console */
-    }
-
-    _resolverFailureHandler(error) {
-        /* eslint-disable no-console */
-        console.error(
-            '@resolver caught error',
-            error,
-            JSON.stringify(error),
-            error.stack
-        )
-        throw error
-        /* eslint-enable no-console */
-    }
-
-    _unsupportedAuthHandler(auth) {
-        /* eslint-disable no-console */
-        console.error(
-            'Auth type ' +
-            auth.constructor.name +
-            ' is not supported in Paw'
-        )
-        /* eslint-enable no-console */
-    }
-
-    _invalidJSONHandler() {
-        /* eslint-disable no-console */
-        console.error(
-            'body was set to JSON, but we couldn\'t parse it'
-        )
-        /* eslint-enable no-console */
-    }
-
-    /*
-      @params:
-        - context
-        - string
-    */
-    createRequestContextFromString(context, string) {
-        return this.createRequestContexts(
-            context,
-            [ { content: string } ],
-            {}
-        )[0]
-    }
-
-    importString(context, string) {
-        let items = [
-            {
-                content: string,
-                url: 'http://localhost/'
-            }
-        ]
-
-        return this.import(context, items)
-    }
-
-    /*
-      @params:
-        - context
-        - items
-        - options
-    */
-    createRequestContexts() {
-        throw new Error('BaseImporter is an abstract class')
-    }
-
-    import(context, items, options) {
-        this.options = (options || {}).inputs || {}
-        /*
-        this.options.jsfInBody = true
-        this.options.jsfInEnv = true
-        this.options.jsfInPath = true
-        this.options.jsfInQuery = true
-        this.options.jsfInHeaders = true
-        */
-
-        this.context = context
-
-        let parsePromiseOrResult = this.createRequestContexts(
-            context,
-            items,
-            options
-        )
-
-        if (typeof parsePromiseOrResult.then !== 'function') {
-            let value = parsePromiseOrResult
-            parsePromiseOrResult = new Promise((resolve) => {
-                resolve(value)
-            })
-        }
-
-
-        let environment = new PawEnvironment()
-        let resolver = new ContextResolver(environment)
-
-        let importPromise = parsePromiseOrResult.then((requestContexts) => {
-            let promises = []
-            for (let env of requestContexts) {
-                promises.push(
-                    this._importContext(
-                        resolver,
-                        env.context,
-                        env.items[0],
-                        options
-                    )
-                )
-            }
-
-
-            return Promise.all(promises).then(() => {
-                return true
-            }, () => {
-                return false
-            })
-        }, ::this._parserErrorHandler).catch(::this._parserFailureHandler)
-
-        return importPromise
-    }
-
-    _importContext(resolver, reqContext, _item, options) {
-        let name = ((_item || {}).file || {}).name || null
-        if (name) {
-            name = name.replace(/\.[^.]*$/, '')
-        }
-        this.currentEnvironmentDomainName = name
-        if (!(reqContext instanceof Context)) {
-            throw new Error(
-                'createRequestContext ' +
-                'did not return an instance of RequestContext'
-            )
-        }
-
-        let item = new Item(_item)
-
-
-        return resolver.resolveAll(
-            item,
-            reqContext,
-            new ResolverOptions({ resolve: { local: false } })
-        ).then(context => {
-            try {
-                this._importPawRequests(
-                    context,
-                    _item,
-                    options
-                )
-                if (options && options.order) {
-                    options.order += 1
-                }
-
-                return true
-            }
-            catch (e) {
-                this._serializerFailureHandler(e)
-            }
-        }, this._resolverErrorHandler).catch(this._resolverFailureHandler)
-    }
-
-    serialize(requestContext, opts = null, item, options) {
-        return this._importPawRequests(requestContext, item, options)
-    }
-
-    _importPawRequests(requestContext, item, options) {
-        const requests = requestContext.get('requests')
-        const group = requestContext.get('group')
-        const references = requestContext.get('references')
-
-        this._importReferences(references)
-
-        const schema = requestContext.get('schema')
-
-        if (group.get('children').size === 0) {
-            return
-        }
-
-        let parent = null
-        let name
-        if (group.get('name')) {
-            name = group.get('name')
-        }
-        else if (item && item.file) {
-            name = item.file.name
-        }
-        else if (item && item.url) {
-            name = item.url
-        }
-
-        if (requests.size > 1) {
-            parent = this.context.createRequestGroup(name)
-        }
-
-        if (options && options.parent && parent) {
-            options.parent.appendChild(parent)
-            if (
-                options &&
-                options.order !== null &&
-                typeof options.order !== 'undefined'
-            ) {
-                parent.order = options.order
-            }
-        }
-
-
-        let manageRequestGroups = (current, parentGroup) => {
-            if (
-                !parentGroup ||
-                current === parentGroup.name ||
-                current === ''
-            ) {
-                return parentGroup
-            }
-            let pawGroup = this.context.createRequestGroup(current)
-            parentGroup.appendChild(pawGroup)
-            return pawGroup
-        }
-
-        this._applyFuncOverGroupTree(
-            requests,
-            group,
-            (request, requestParent) => {
-                ::this._importPawRequest(
-                    options,
-                    requestParent,
-                    request,
-                    schema
-                )
-            },
-            manageRequestGroups,
-            parent
-        )
-    }
-
-    _importReferences(references) {
-        let environmentDomain = this._getEnvironmentDomain()
-
-        let variablesDict = {}
-
-        let environments = references.keySeq()
-
-        for (let env of environments) {
-            let container = references.get(env)
-
-            let pawEnv = this._getEnvironment(
-                environmentDomain,
-                container.get('name') || container.get('id') || env
-            )
-            let uris = container.get('cache').keySeq()
-            for (let uri of uris) {
-                let reference = container.resolve(uri)
-
-                /*
-                    LateResolutionReference defines uris in the form of
-                    `#/x-postman/~1users~1{{userId}}` as this uri uniquely
-                    identifies the char sequence `/user/{{userId}}`. These
-                    uris are necessary to prevent collision when parsing,
-                    however, since Paw has DynamicStrings it can display and
-                    store these references nicely, so we should not save them
-                    as environment variables, but directly as dynamic string
-                    where they are used.
-
-                    .slice(12) is there to remove `#/x-postman/` from the
-                    string before testing to see if it is a root
-                    LateResolutionReference or not.
-                */
-                if (!(reference instanceof LateResolutionReference) ||
-                    (reference.get('uri') || '').slice(12).match(/^{{[^{}]+}}$/)
-                ) {
-                    let content = this._setReference(reference)
-                    let ds
-                    if (content instanceof DynamicString) {
-                        ds = content
-                    }
-                    else {
-                        ds = new DynamicString(content || '')
-                    }
-
-                    let key = reference.get('relative')
-                    if (reference instanceof LateResolutionReference) {
-                        // remove '#/x-postman/' and {{ }}
-                        key = key.slice(12)
-                        key = key.slice(2, key.length - 2)
-                    }
-
-                    if (
-                        key !== null &&
-                        key !== '' &&
-                        typeof key !== 'undefined'
-                    ) {
-                        variablesDict[key] = ds
-                    }
-                }
-            }
-
-            pawEnv.setVariablesValues(variablesDict)
-        }
-    }
-
-    _setReference(reference) {
-        if (reference instanceof JSONSchemaReference) {
-            return this._setJSONSchemaReference(reference)
-        }
-        else if (reference instanceof LateResolutionReference) {
-            return this._setLateResolutionReference(reference)
-        }
-        else if (reference.get('value') === null) {
-            return this._setExoticReference(reference)
-        }
-        else {
-            return this._setSimpleReference(reference)
-        }
-    }
-
-    _setJSONSchemaReference(reference) {
-        let dv
-        if (this.options.jsfInEnv) {
-            dv = new DynamicValue(
-                'com.luckymarmot.PawExtensions.JSONSchemaFakerDynamicValue',
-                {
-                    schema: JSON.stringify(reference.toJSONSchema() || {})
-                }
-            )
-        }
-        else {
-            dv = ''
-        }
-        return dv
-    }
-
-    _setLateResolutionReference(reference) {
-        // slice(12) because we slice out '#/x-postman/' from the uri
-        let ref = (reference.get('relative') || reference.get('uri') || '')
-            .slice(12)
-        let match = ref.match(/({{[^{}]*}})/g)
-        if (match && ref !== match[0]) {
-            let components = []
-            let baseIndex = 0
-            let re = /({{[^{}]+}})/g
-            let m
-            while ((m = re.exec(ref)) !== null) {
-                let index = m.index
-                if (baseIndex !== index) {
-                    components.push(this._unescapeURIFragment(
-                        ref.slice(baseIndex, index)
-                    ))
-                }
-
-                baseIndex = index + m[0].length
-
-                let envVariable = this._getEnvironmentVariable(
-                    m[0].slice(2, m[0].length - 2)
-                )
-
-                let dv = new DynamicValue(
-                    'com.luckymarmot.EnvironmentVariableDynamicValue',
-                    {
-                        environmentVariable: envVariable.id
-                    }
-                )
-
-                components.push(dv)
-            }
-
-            components.push(this._unescapeURIFragment(
-                ref.slice(baseIndex, ref.length)
-            ))
-
-            return new DynamicString(...components)
-        }
-        else {
-            let value = reference.get('value')
-            if (typeof value !== 'string' || value === '') {
-                return ref.slice(2, ref.length - 2)
-            }
-            return value
-        }
-    }
-
-    _unescapeURIFragment(uriFragment) {
-        return uriFragment.replace(/~1/g, '/').replace(/~0/g, '~')
-    }
-
-    _setExoticReference(reference) {
-        if (reference.get('uri').match('://')) {
-            return new DynamicValue(
-                'com.luckymarmot.PawExtensions.RemoteFileDynamicValue',
-                {
-                    url: reference.get('uri')
-                }
-            )
-        }
-        else {
-            return new DynamicValue(
-                'com.luckymarmot.FileContentDynamicValue', {
-                    filePath: reference.get('uri')
-                }
-            )
-        }
-    }
-
-    _setSimpleReference(reference) {
-        return reference.get('value')
-    }
-
-    _getEnvironmentDomain() {
-        let env = this.context.getEnvironmentDomainByName(
-            this.currentEnvironmentDomainName ||
-            this.ENVIRONMENT_DOMAIN_NAME
-        )
-        if (typeof env === 'undefined') {
-            env = this.context
-                .createEnvironmentDomain(
-                    this.currentEnvironmentDomainName ||
-                    this.ENVIRONMENT_DOMAIN_NAME
-                )
-        }
-        return env
-    }
-
-    _getEnvironment(domain, environmentName = 'Default Environment') {
-        let env = domain.getEnvironmentByName(environmentName)
-        if (typeof env === 'undefined') {
-            env = domain.createEnvironment(environmentName)
-        }
-        return env
-    }
-
-    _getEnvironmentVariable(name) {
-        let domain = this._getEnvironmentDomain()
-        let variable = domain.getVariableByName(name)
-        if (typeof variable === 'undefined') {
-            let env = this._getEnvironment(domain)
-            let varD = {}
-            varD[name] = ''
-            env.setVariablesValues(varD)
-            variable = domain.getVariableByName(name)
-        }
-        return variable
-    }
-
-    _importPawRequest(options, parent, request, schema) {
-        let container = request.get('parameters')
-        let bodies = request.get('bodies')
-        const auths = request.get('auths')
-        const timeout = request.get('timeout')
-
-        let contentType
-        if (bodies.size > 0) {
-            let body
-            body = bodies.get(0)
-            contentType = this._extractContentTypeFromBody(body)
-            container = container.filter(body.get('constraints'))
-        }
-
-        // url + method
-        let pawRequest = this._createPawRequest(request, container)
-
-        pawRequest.description = request.get('description')
-
-        // headers
-        pawRequest = this._setHeaders(pawRequest, container)
-
-        // auth
-        pawRequest = ::this._setAuth(pawRequest, auths)
-
-        // body
-        pawRequest = this._setBody(
-            pawRequest,
-            contentType,
-            container,
-            schema
-        )
-
-        // timeout
-        if (timeout) {
-            pawRequest.timeout = timeout * 1000
-        }
-
-        if (parent) {
-            parent.appendChild(pawRequest)
-        }
-        else if (options && options.parent) {
-            options.parent.appendChild(pawRequest)
-        }
-
-        // order
-        if (options && options.order) {
-            pawRequest.order = options.order
-        }
-
-        return pawRequest
-    }
-
-    _applyFuncOverGroupTree(requests, group, leafFunc, nodeFunc, pawGroup) {
-        let calls = []
-        let currentPawGroup = nodeFunc(group.get('name') || '', pawGroup)
-        group.get('children').forEach((child) => {
-            if (typeof child === 'string' || typeof child === 'number') {
-                const request = requests.get(child)
-                if (request) {
-                    calls.push(leafFunc(request, currentPawGroup))
-                }
-            }
-            else {
-                calls = calls.concat(
-                    this._applyFuncOverGroupTree(
-                        requests,
-                        child,
-                        leafFunc,
-                        nodeFunc,
-                        currentPawGroup
-                    )
-                )
-            }
-        })
-        return calls
-    }
-
-    _createPawRequest(request, container) {
-        let url = ::this._generateUrl(
-            request.get('url'),
-            container.get('queries'),
-            request.get('auths')
-        )
-        return this.context.createRequest(
-            request.get('name'),
-            (request.get('method') || 'get').toUpperCase(),
-            url,
-        )
-    }
-
-    _generateUrl(url, queries, auths) {
-        let protocol = this._formatProtocolParam(url.get('protocol'))
-        let host = this._toDynamicString(url.get('host'), true, 'host')
-        let path = this._toDynamicString(url.get('pathname'), true, 'pathname')
-        let hash = this._toDynamicString(url.get('hash'), true, 'url')
-
-        if (protocol.length > 0) {
-            protocol.appendString(':')
-        }
-
-        if (protocol.length > 0 || host.length > 0) {
-            protocol.appendString('//')
-        }
-
-        let _url = new DynamicString(
-            ...protocol.components,
-            ...host.components,
-            ...path.components
-        )
-
-        let queryParams = (queries || []).concat(
-            this._extractQueryParamsFromAuth(auths)
-        )
-        if (queryParams.size > 0) {
-            _url.appendString('?')
-            let _params = queryParams.reduce(
-                (params, keyValue) => {
-                    let dynKey = this._toDynamicString(
-                        keyValue.get('key'), true, 'query'
-                    ).components.map((component) => {
-                        if (typeof component === 'string') {
-                            return encodeURI(component)
-                        }
-                        return component
-                    })
-                    let dynValue = this._toDynamicString(
-                        keyValue, true, 'query'
-                    ).components.map((component) => {
-                        if (typeof component === 'string') {
-                            return encodeURI(component)
-                        }
-                        return component
-                    })
-                    let param = []
-                    if (params.length !== 0) {
-                        param.push('&')
-                    }
-                    param = param.concat(dynKey)
-                    param.push('=')
-                    param = param.concat(dynValue)
-                    return params.concat(param)
-                },
-                []
-            )
-            _url = new DynamicString(
-                ..._url.components,
-                ..._params,
-                ...hash.components
-            )
-        }
-        else {
-            _url = new DynamicString(
-                ..._url.components,
-                ...hash.components
-            )
-        }
-
-        return _url
-    }
-
-    _formatProtocolParam(protocol) {
-        if (this._useJSF('protocol')) {
-            return this._toDynamicString(protocol, true, 'protocol')
-        }
-
-        let schema = protocol.getJSONSchema(false)
-
-        if (schema && schema.enum.indexOf('https') >= 0) {
-            return new DynamicString('https')
-        }
-        else if (schema && schema.enum.indexOf('wss') >= 0) {
-            return new DynamicString('wss')
-        }
-        else {
-            return this._toDynamicString(protocol, true, 'protocol')
-        }
-    }
-
-    _extractQueryParamsFromAuth(auths) {
-        return (auths || []).filter((auth) => {
-            return auth instanceof ApiKeyAuth && auth.get('in') === 'query'
-        }).map((auth) => {
-            return new Parameter({
-                key: auth.get('name'),
-                value: auth.get('key')
-            })
-        }).toArray()
-    }
-
-    _setHeaders(pawReq, container) {
-        let headers = container.getHeadersSet()
-        headers.forEach((param) => {
-            pawReq.setHeader(
-                this._toDynamicString(param.get('key'), true, 'headers'),
-                this._toDynamicString(param, true, 'headers')
-            )
-        })
-        return pawReq
-    }
-
-    _setBasicAuth(auth) {
-        return new DynamicValue(
-            'com.luckymarmot.BasicAuthDynamicValue',
-            {
-                username: this._toDynamicString(
-                    auth.get('username') || '', true, 'auth'
-                ),
-                password: this._toDynamicString(
-                    auth.get('password') || '', true, 'auth'
-                )
-            }
-        )
-    }
-
-    _setDigestAuth(auth) {
-        return new DynamicValue(
-            'com.luckymarmot.PawExtensions.DigestAuthDynamicValue',
-            {
-                username: this._toDynamicString(
-                    auth.get('username'), true, 'auth'
-                ),
-                password: this._toDynamicString(
-                    auth.get('password'), true, 'auth'
-                )
-            }
-        )
-    }
-
-    _setOAuth1Auth(auth) {
-        return new DynamicValue(
-            'com.luckymarmot.OAuth1HeaderDynamicValue',
-            {
-                callback: this._toDynamicString(
-                    auth.get('callback') || '', true, 'auth'
-                ),
-                consumerKey: this._toDynamicString(
-                    auth.get('consumerKey') || '', true, 'auth'
-                ),
-                consumerSecret: this._toDynamicString(
-                    auth.get('consumerSecret') || '', true, 'auth'
-                ),
-                tokenSecret: this._toDynamicString(
-                    auth.get('tokenSecret') || '', true, 'auth'
-                ),
-                algorithm: auth.get('algorithm') || '',
-                nonce: this._toDynamicString(
-                    auth.get('nonce') || '', true, 'auth'
-                ),
-                additionalParamaters: auth
-                    .get('additionalParamaters') || '',
-                timestamp: this._toDynamicString(
-                    auth.get('timestamp') || '', true, 'auth'
-                ),
-                token: this._toDynamicString(
-                    auth.get('token') || '', true, 'auth'
-                )
-            }
-        )
-    }
-
-    _setOAuth2Auth(auth) {
-        const grantMap = {
-            accessCode: 0,
-            implicit: 1,
-            application: 2,
-            password: 3
-        }
-        return new DynamicValue(
-            'com.luckymarmot.OAuth2DynamicValue',
-            {
-                grantType: grantMap[auth.get('flow')] || 0,
-                authorizationUrl: this._toDynamicString(
-                    auth.get('authorizationUrl') || '', true, 'auth'
-                ),
-                accessTokenUrl: this._toDynamicString(
-                    auth.get('tokenUrl') || '', true, 'auth'
-                ),
-                scope: (auth.get('scopes') || []).join(' ')
-            }
-        )
-    }
-
-    _setAWSSig4Auth(auth) {
-        return new DynamicValue(
-            'com.shigeoka.PawExtensions.AWSSignature4DynamicValue',
-            {
-                key: this._toDynamicString(
-                    auth.get('key') || '', true, 'auth'
-                ),
-                secret: this._toDynamicString(
-                    auth.get('secret') || '', true, 'auth'
-                ),
-                region: this._toDynamicString(
-                    auth.get('region') || '', true, 'auth'
-                ),
-                service: this._toDynamicString(
-                    auth.get('service') || '', true, 'auth'
-                )
-            }
-        )
-    }
-
-    _setHawkAuth(auth) {
-        return new DynamicValue(
-            'uk.co.jalada.PawExtensions.HawkDynamicValue',
-            {
-                key: this._toDynamicString(
-                    auth.get('key') || '', true, 'auth'
-                ),
-                id: this._toDynamicString(
-                    auth.get('id') || '', true, 'auth'
-                ),
-                algorithm: this._toDynamicString(
-                    auth.get('algorithm') || '', true, 'auth'
-                )
-            }
-        )
-    }
-
-    _setAuth(pawReq, auths) {
-        const authTypeMap = {
-            BasicAuth: ::this._setBasicAuth,
-            DigestAuth: ::this._setDigestAuth,
-            OAuth1Auth: ::this._setOAuth1Auth,
-            OAuth2Auth: ::this._setOAuth2Auth,
-            AWSSig4Auth: ::this._setAWSSig4Auth,
-            HawkAuth: ::this._setHawkAuth
-        }
-
-        for (let auth of auths) {
-            let rule = authTypeMap[auth.constructor.name]
-
-            if (rule) {
-                const dv = rule(auth)
-                pawReq.setHeader('Authorization', new DynamicString(dv))
-            }
-            else if (auth instanceof ApiKeyAuth) {
-                if (auth.get('in') === 'header') {
-                    pawReq.setHeader(
-                        this._toDynamicString(auth.get('name'), true, 'auth'),
-                        this._toDynamicString(auth.get('key'), true, 'auth')
-                    )
-                }
-            }
-            else {
-                this._unsupportedAuthHandler(auth)
-            }
-        }
-        return pawReq
-    }
-
-    _setBody(pawReq, contentType, container) {
-        let body = container.get('body')
-
-        const bodyRules = {
-            'multipart/form-data':
-                ::this._setFormDataBody,
-            'application/x-www-form-urlencoded':
-                ::this._setUrlEncodedBody,
-            'application/json':
-                ::this._setJSONBody
-        }
-
-        let _pawReq = pawReq
-
-        const rule = bodyRules[contentType]
-        if (rule) {
-            _pawReq = rule(pawReq, body)
-        }
-        else if (body.size > 1) {
-            _pawReq = this._setUrlEncodedBody(pawReq, body)
-        }
-        else {
-            this._setPlainBody(pawReq, body)
-        }
-
-        return _pawReq
-    }
-
-    _extractContentTypeFromBody(body) {
-        let constraints = body.get('constraints')
-        let contentType = null
-        constraints.forEach(param => {
-            if (param.get('key') === 'Content-Type') {
-                contentType = param.get('value')
-            }
-        })
-
-        return contentType
-    }
-
-    _setFormDataBody(pawReq, body) {
-        pawReq.setHeader(
-            'Content-Type',
-            'multipart/form-data'
-        )
-        const keyValues = body.map(param => {
-            let key = this._toDynamicString(
-                param.get('key'), true, 'body'
-            )
-            let value = this._toDynamicString(
-                param.get, true, 'body'
-            )
-            return [ key, value, true ]
-        }).toArray()
-        const dv = new DynamicValue(
-            'com.luckymarmot.BodyMultipartFormDataDynamicValue', {
-                keyValues: keyValues
-            }
-        )
-        pawReq.body = new DynamicString(dv)
-        return pawReq
-    }
-
-    _setUrlEncodedBody(pawReq, body) {
-        pawReq.setHeader(
-            'Content-Type',
-            'application/x-www-form-urlencoded'
-        )
-        const keyValues = body.map(param => {
-            let key = this._toDynamicString(
-                param.get('key'), true, 'body'
-            )
-            let value = this._toDynamicString(
-                param, true, 'body'
-            )
-            return [ key, value, true ]
-        }).toArray()
-        const dv = new DynamicValue(
-            'com.luckymarmot.BodyFormKeyValueDynamicValue', {
-                keyValues: keyValues
-            }
-        )
-        pawReq.body = new DynamicString(dv)
-        return pawReq
-    }
-
-    _setPlainBody(pawReq, body) {
-        if (body.size > 0) {
-            let param = body.get(0)
-            pawReq.body = this._toDynamicString(param, true, 'body')
-        }
-        else {
-            pawReq.body = ''
-        }
-        return pawReq
-    }
-
-    _setJSONBody(pawReq, body) {
-        pawReq.setHeader(
-            'Content-Type',
-            'application/json'
-        )
-
-        if (body.size === 0) {
-            return pawReq
-        }
-
-        let param = body.get(0)
-        let content = this._toDynamicString(param, true, 'body')
-
-        let component = content.getComponentAtIndex(0)
-
-        if (
-            content.length === 1 &&
-            typeof component === 'string'
-        ) {
-            try {
-                pawReq.jsonBody = JSON.parse(component)
-            }
-            catch (e) {
-                this._invalidJSONHandler()
-                pawReq.body = content
-            }
-        }
-        else {
-            pawReq.body = content
-        }
-
-        return pawReq
-    }
-
-    _toDynamicString(string, defaultToEmpty, source) {
-        if (!string) {
-            if (defaultToEmpty) {
-                return new DynamicString('')
-            }
-            return null
-        }
-
-        let envComponents = []
-        if (string instanceof Parameter) {
-            if (string.get('type') === 'reference') {
-                envComponents = this._castReferenceToDynamicString(
-                    string.get('value')
-                ).components
-            }
-            else {
-                envComponents = this._castParameterToDynamicString(
-                    string, source
-                ).components
-            }
-        }
-        else if (typeof string === 'string') {
-            envComponents.push(string)
-        }
-
-        let components = []
-
-        let splitManager = (_isRegular) => {
-            let isRegular = _isRegular
-            return (split) => {
-                if (isRegular) {
-                    components.push(split)
-                }
-                else {
-                    components.push(
-                        this._escapeSequenceDynamicValue(split)
-                    )
-                }
-                isRegular = !isRegular
-            }
-        }
-
-        for (let component of envComponents) {
-            if (typeof component !== 'string') {
-                components.push(component)
-            }
-            else {
-                // split around special characters
-                const re = /([^\x00-\x1f]+)|([\x00-\x1f]+)/gm
-                let splits = component.match(re)
-                if (splits.length > 0 && splits.length < 50) {
-                    let isRegular = /([^\x00-\x1f]+)/.test(splits[0])
-                    let splitMapper = splitManager(isRegular)
-                    splits.forEach(splitMapper)
-                }
-                else {
-                    components.push(component)
-                }
-            }
-        }
-
-        return new DynamicString(...components)
-    }
-
-    _convertCharToHex(char) {
-        let hexChar = char.charCodeAt(0).toString(16)
-        if (hexChar.length === 1) {
-            hexChar = '0' + hexChar
-        }
-        return hexChar
-    }
-
-    _escapeCharSequence(seq) {
-        const escapedChars = {
-            '\n': '\\n',
-            '\r': '\\r',
-            '\t': '\\t'
-        }
-
-        let sequence = []
-        for (let char of seq) {
-            sequence.push(
-                escapedChars[char] ?
-                    escapedChars[char] :
-                    '\\x' + this._convertCharToHex(char)
-            )
-        }
-        return sequence.join('')
-    }
-
-    _escapeSequenceDynamicValue(seq) {
-        let escapeSequence = this._escapeCharSequence(seq)
-        return new DynamicValue('com.luckymarmot.EscapeSequenceDynamicValue', {
-            escapeSequence: escapeSequence
-        })
-    }
-
-    _castReferenceToDynamicString(reference) {
-        let dv = this._extractReferenceComponent(reference)
-        if (dv instanceof DynamicString) {
-            return dv
-        }
-        return new DynamicString(dv)
-    }
-
-    _castParameterToDynamicString(param, source) {
-        let schema = param.getJSONSchema(false)
-
-        let components = []
-        if (schema['x-sequence']) {
-            for (let item of schema['x-sequence']) {
-                if (item['x-title']) {
-                    if (schema.enum && schema.enum.length === 1) {
-                        components.push(param.generate())
-                    }
-                    else if (this._useJSF(source)) {
-                        let dv = new DynamicValue(
-                            'com.luckymarmot.PawExtensions' +
-                            '.JSONSchemaFakerDynamicValue',
-                            {
-                                schema: JSON.stringify(item)
-                            }
-                        )
-                        components.push(dv)
-                    }
-                    else {
-                        components.push(param.generate(false, item))
-                    }
-                }
-                else {
-                    components.push(param.generate(false, item))
-                }
-            }
-        }
-        else if (schema.enum && schema.enum.length === 1) {
-            let generated = param.generate(false, schema)
-            if (generated === null) {
-                generated = ''
-            }
-            components.push(generated)
-        }
-        else if (this._useJSF(source)) {
-            let dv = new DynamicValue(
-                'com.luckymarmot.PawExtensions' +
-                '.JSONSchemaFakerDynamicValue',
-                {
-                    schema: JSON.stringify(schema)
-                }
-            )
-            components.push(dv)
-        }
-        else {
-            let generated = param.generate(false, schema)
-            if (generated === null) {
-                generated = ''
-            }
-            components.push(generated)
-        }
-
-        return new DynamicString(...components)
-    }
-
-    _useJSF(source) {
-        let validSources = [
-            'protocol',
-            'host',
-            'pathname',
-            'query',
-            'body',
-            'headers'
-        ]
-
-        if (validSources.indexOf(source) < 0) {
-            return true
-        }
-
-        /* eslint-disable no-extra-parens */
-        return (source === 'protocol' && this.options.jsfInProtocol) ||
-            (source === 'host' && this.options.jsfInHost) ||
-            (source === 'pathname' && this.options.jsfInPath) ||
-            (source === 'query' && this.options.jsfInQuery) ||
-            (source === 'body' && this.options.jsfInBody) ||
-            (source === 'headers' && this.options.jsfInHeaders)
-        /* eslint-enable no-extra-parens */
-    }
-
-    _extractReferenceComponent(component) {
-        if (typeof component === 'string') {
-            return component
-        }
-
-        if (component instanceof Reference) {
-            /*
-                `!component.get('relative')` is not related to the explanation
-                below.
-
-                If the component is a LateResolutionReference, we prefer to
-                have directly access to the DynamicString that represents it,
-                than having a reference to it stored in the environment domain.
-            */
-            if (
-                !component.get('relative')
-            ) {
-                return this._setReference(component)
-            }
-
-            if (component instanceof LateResolutionReference) {
-                let match = (component.get('uri') || '')
-                .slice(12)
-                .match(/^{{[^{}]+}}$/)
-                if (!match) {
-                    return this._setReference(component)
-                }
-                else {
-                    let envVariable = this._getEnvironmentVariable(
-                        match[0].slice(2, match[0].length - 2)
-                    )
-                    return new DynamicValue(
-                        'com.luckymarmot.EnvironmentVariableDynamicValue',
-                        {
-                            environmentVariable: envVariable.id
-                        }
-                    )
-                }
-            }
-
-            let envVariable = this._getEnvironmentVariable(
-                component.get('relative')
-            )
-            return new DynamicValue(
-                'com.luckymarmot.EnvironmentVariableDynamicValue',
-                {
-                    environmentVariable: envVariable.id
-                }
-            )
-        }
-
-        return null
-    }
+  serialize(api, options) {
+    return methods.serialize(api, options)
+  }
 }
+
+/**
+ * wraps a DynamicValue in a DynamicString
+ * @param {DynamicValue} dv: the dv to wrapDV
+ * @returns {DynamicString} the corresponding DynamicString
+ */
+methods.wrapDV = (dv) => new DynamicString(dv)
+
+/**
+ * creates a JSON DynamicValue to wrap a JSON object.
+ * @param {Object} json: the json object to wrap in a DV
+ * @returns {DynamicValue} the corresponding JSON DynamicValue
+ */
+methods.createJSONDV = (json) => new DynamicValue('com.luckymarmot.JSONDynamicValue', {
+  json: JSON.stringify(json)
+})
+
+/**
+ * extracts the title from an Api Record.
+ * @param {Api} api: the api to get the name of.
+ * @returns {string} the title of the api, or a default value if the api does not have any title
+ */
+methods.getTitleFromApi = (api) => api.getIn([ 'info', 'title' ]) || 'Imports'
+
+/**
+ * creates a standard environment domain in which to store standard values, such as shared schemas,
+ * parameters, endpoints and authentication methods.
+ * @param {Context} context: the paw context in which to create the environment domain.
+ * @param {Api} api: the api to import in Paw.
+ * @returns {EnvironmentDomain} the newly created environment domain
+ */
+methods.createStandardEnvironmentDomain = (context, api) => {
+  const title = methods.getTitleFromApi(api)
+  return context.createEnvironmentDomain(title)
+}
+
+// assumes correctly formatted api
+/**
+ * calculates the size of the standard environment domain. This is used to determine whether we
+ * should create a domain or not.
+ * @param {Api} api: the api to import in Paw.
+ * @returns {integer} the number of values that would be stored in the environment domain.
+ */
+methods.getStandardEnvironmentDomainSize = (api) => {
+  const constraints = api.getIn([ 'store', 'constraint' ])
+  const endpoints = api.getIn([ 'store', 'endpoint' ])
+  const auths = api.getIn([ 'store', 'auth' ])
+  const parameters = api.getIn([ 'store', 'parameter' ])
+  // const responses = api.getIn([ 'store', 'response' ])
+
+  const size = constraints.size +
+    endpoints.size +
+    auths.size +
+    parameters.size
+    // +
+    // responses.size
+
+  return size
+}
+
+/**
+ * tests whether an Api needs a standard environment domain to be fully imported in Paw
+ * @param {Api} api: the api to import in Paw
+ * @returns {boolean} whether a standard environment domain is needed.
+ */
+methods.needsStandardEnvironmentDomain = (api) => {
+  const size = methods.getStandardEnvironmentDomainSize(api)
+
+  return size > 0
+}
+
+/**
+ * creates a Variable environment domain. This is used to import Variables with multiple contexts
+ * into Paw.
+ * @param {Context} context: the Paw context in which to create the environment domain
+ * @param {Api} api: the api to import in Paw.
+ * @returns {EnvironmentDomain} the newly created environment domain.
+ *
+ * NOTE: It needs to be separate from the standard domain because values imported in the standard
+ * domain only have a single value possible, whereas variables can have multiple values whose
+ * evaluation depend on the name of the context (Think switching between Postman Environments)
+ */
+methods.createVariableEnvironmentDomain = (context, api) => {
+  const title = methods.getTitleFromApi(api)
+  const domainName = 'Vars - ' + title
+  return context.createEnvironmentDomain(domainName)
+}
+
+/**
+ * adds a Constraint as JSON DV to a domain.
+ * @param {EnvironmentDomain} domain: the domain to add the constraint to.
+ * @param {Environment} environment: the domain environment for which this value is applicable.
+ * @param {Constraint} constraint: the constraint to add.
+ * @param {string} key: the name of the constraint
+ * @returns {EnvironmentVariable} the newly created environment variable
+ */
+methods.addConstraintToDomain = (domain, environment, constraint, key) => {
+  const variable = domain.createEnvironmentVariable(key)
+  const schema = constraint.toJSONSchema()
+  const dv = methods.createJSONDV(schema)
+  const ds = methods.wrapDV(dv)
+  variable.setValue(ds, environment)
+
+  return variable
+}
+
+/**
+ * adds all constraints of a Constraint TypedStore into a domain.
+ * @param {EnvironmentDomain} domain: the domain to add the constraints to.
+ * @param {Environment} environment: the domain environment for which the constraints are
+ * applicable.
+ * @param {Api} api: the api to get the TypedStore from.
+ * @returns {TypedStore}: a TypedStore of EnvironmentVariables representing constraints.
+ */
+methods.addConstraintsToDomain = (domain, environment, api) => {
+  const constraints = api.getIn([ 'store', 'constraint' ])
+  const addConstraint = currify(methods.addConstraintToDomain, domain, environment)
+
+  return constraints.map(addConstraint)
+}
+
+/**
+ * removes ':' at the end of a protocol string
+ * @param {string} protocol: the protocol to trim
+ * @returns {string} the trimmed string
+ */
+methods.removeDotsFromProtocol = (protocol) => {
+  if (protocol[protocol.length - 1] === ':') {
+    return protocol.slice(0, protocol.length - 1)
+  }
+
+  return protocol
+}
+
+/**
+ * converts a protocol string into a Record Parameter.
+ * @param {string} protocol: the protocol to convert
+ * @param {integer} index: the index of the protocol in the parent array. This is used to set the
+ * first Record Parameter as enabled
+ * @return {RecordParameter} the corresponding Record Parameter.
+ */
+methods.convertProtocolIntoRecordParameter = (protocol, index) => {
+  const stripped = methods.removeDotsFromProtocol(protocol)
+  const isEnabled = index === 0
+  return new RecordParameter(stripped, ', ', isEnabled)
+}
+
+/**
+ * converts a protocol URLComponent into a DynamicValue.
+ * @param {URLComponent} protocol: the url component to convert
+ * @returns {DynamicValue} the corresponding DynamicValue.
+ */
+methods.createProtocolDV = (protocol) => {
+  if (!protocol || !protocol.size) {
+    return 'http'
+  }
+
+  if (protocol.size === 1) {
+    return methods.removeDotsFromProtocol(protocol.get(0))
+  }
+
+  return new DynamicValue('me.elliotchance.MultiSelectorDynamicValue', {
+    choices: protocol.map(methods.convertProtocolIntoRecordParameter).toJS(),
+    separator: ','
+  })
+}
+
+// TODO save parameters as document parameters when these finally exist in Paw
+/**
+ * converts a urlComponent into a DynamicString or a standard string as a fallback.
+ * @param {URLComponent} urlComponent: the urlComponent to convert
+ * @return {string|DynamicString} the converted value
+ */
+methods.convertURLComponentToDynamicString = (urlComponent) => {
+  if (!urlComponent) {
+    return ''
+  }
+
+  return urlComponent.generate(List([ '{', '}' ]))
+}
+
+/**
+ * converts an endpoint into a DynamicString.
+ * @param {URL} endpoint: the endpoint to convert
+ * @returns {DynamicString} the corresponding DynamicString
+ */
+methods.createEndpointDynamicString = (endpoint) => {
+  const protocol = methods.createProtocolDV(endpoint.get('protocol'))
+  const slashes = '://'
+  const hostname = methods.convertURLComponentToDynamicString(endpoint.get('hostname'))
+  const port = methods.convertURLComponentToDynamicString(endpoint.get('port'))
+  const portDots = port ? ':' : ''
+  const pathname = methods.convertURLComponentToDynamicString(endpoint.get('pathname'))
+
+  return new DynamicString(protocol, slashes, hostname, portDots, port, pathname)
+}
+
+/**
+ * adds an endpoint to a domain, as DynamicStrings
+ * @param {EnvironmentDomain} domain: the domain to add the endpoint to.
+ * @param {Environment} environment: the environment in which this endpoint value is applicable.
+ * @param {URL} endpoint: the endpoint to add to the domain
+ * @param {string} key: the name of the endpoint
+ * @returns {EnvironmentVariable} the newly created environment variable.
+ */
+methods.addEndpointToDomain = (domain, environment, endpoint, key) => {
+  const variable = domain.createEnvironmentVariable(key)
+  const ds = methods.createEndpointDynamicString(endpoint)
+  variable.setValue(ds, environment)
+
+  return variable
+}
+
+/**
+ * adds all endpoints from an Endpoint TypedStore to a domain.
+ * @param {EnvironmentDomain} domain: the domain to add the endpoints to.
+ * @param {Environment} environment: the environment in which these endpoints are applicable.
+ * @param {Api} api: the api from which to get the Endpoint TypedStore.
+ * @returns {TypedStore} a TypedStore of EnvironmentVariables representing endpoints.
+ */
+methods.addEndpointsToDomain = (domain, environment, api) => {
+  const endpoints = api.getIn([ 'store', 'endpoint' ])
+  const addEndpoint = currify(methods.addEndpointToDomain, domain, environment)
+
+  return endpoints.map(addEndpoint)
+}
+
+/**
+ * adds a Parameter to a domain, as a DynamicString
+ * @param {EnvironmentDomain} domain: the domain to add the parameter to.
+ * @param {Environment} environment: the environment in which this parameter value is applicable.
+ * @param {Parameter} parameter: the parameter to add to the domain
+ * @param {string} key: the name of the endpoint
+ * @returns {EnvironmentVariable} the newly created environment variable.
+ */
+methods.addParameterToDomain = (domain, environment, parameter, key) => {
+  const variable = domain.createEnvironmentVariable(key)
+  const schema = parameter.getJSONSchema(false)
+  const dv = methods.createJSONDV(schema)
+  const ds = methods.wrapDV(dv)
+  variable.setValue(ds, environment)
+
+  return variable
+}
+
+/**
+ * adds all parameters from a Parameter TypedStore to a domain.
+ * @param {EnvironmentDomain} domain: the domain to add the parameters to.
+ * @param {Environment} environment: the environment in which these parameters are applicable.
+ * @param {Api} api: the api from which to get the Parameter TypedStore.
+ * @returns {TypedStore} a TypedStore of EnvironmentVariables representing parameters.
+ */
+methods.addParametersToDomain = (domain, environment, api) => {
+  const parameters = api.getIn([ 'store', 'parameter' ])
+  const addParameter = currify(methods.addParameterToDomain, domain, environment)
+
+  return parameters.map(addParameter)
+}
+
+/**
+ * converts a BasicAuth into its corresponding DynamicValue
+ * @param {Auth} auth: the basic auth to convert
+ * @returns {DynamicValue} the corresponding DynamicValue
+ */
+methods.convertBasicAuthIntoDynamicValue = (auth) => {
+  return new DynamicValue('com.luckymarmot.BasicAuthDynamicValue', {
+    username: auth.get('username') || '',
+    password: auth.get('password') || ''
+  })
+}
+
+/**
+ * converts an auth in its corresponding DynamicValue depending on the type of the auth.
+ * @param {Auth} auth: the auth to convert
+ * @returns {DynamicValue|''} the corresponding DynamicValue, or an empty string if there are no
+ * equivalent for this auth in Paw.
+ */
+methods.convertAuthIntoDynamicValue = (auth) => {
+  if (auth instanceof Auth.Basic) {
+    return methods.convertBasicAuthIntoDynamicValue(auth)
+  }
+
+  return ''
+}
+
+/**
+ * adds an Auth to a domain, as a DynamicString
+ * @param {EnvironmentDomain} domain: the domain to add the auth to.
+ * @param {Environment} environment: the environment in which this auth value is applicable.
+ * @param {Auth} auth: the auth to add to the domain
+ * @param {string} key: the name of the auth
+ * @returns {EnvironmentVariable} the newly created environment variable.
+ */
+methods.addAuthToDomain = (domain, environment, auth, key) => {
+  const variable = domain.createEnvironmentVariable(key)
+  const dv = methods.convertAuthIntoDynamicValue(auth)
+  const ds = methods.wrapDV(dv)
+  variable.setValue(ds, environment)
+
+  return variable
+}
+
+/**
+ * adds all auths from an Auth TypedStore to a domain.
+ * @param {EnvironmentDomain} domain: the domain to add the auths to.
+ * @param {Environment} environment: the environment in which these auths are applicable.
+ * @param {Api} api: the api from which to get the Auth TypedStore.
+ * @returns {TypedStore} a TypedStore of EnvironmentVariables representing auths.
+ */
+methods.addAuthsToDomain = (domain, environment, api) => {
+  const auths = api.getIn([ 'store', 'auth' ])
+  const addAuth = currify(methods.addAuthToDomain, domain, environment)
+
+  return auths.map(addAuth)
+}
+
+
+methods.addVariablesToStandardDomain = (context, domain, api) => {
+  const environment = domain.createEnvironment('Default')
+
+  const constraint = methods.addConstraintsToDomain(domain, environment, api)
+  const endpoint = methods.addEndpointsToDomain(domain, environment, api)
+  const parameter = methods.addParametersToDomain(domain, environment, api)
+  const auth = methods.addAuthsToDomain(domain, environment, api)
+
+  return new Store({ constraint, endpoint, parameter, auth })
+}
+
+methods.getVariableEnvironmentDomainSize = (api) => api.getIn([ 'store', 'variable' ]).size
+
+methods.needsVariableEnvironmentDomain = (api) => {
+  const size = methods.getVariableEnvironmentDomainSize(api)
+
+  return size
+}
+
+methods.updateEnvironmentVariableWithEnvironmentValue = (domain, variable, value, envName) => {
+  let environment = domain.getEnvironmentByName(envName)
+  if (!environment) {
+    environment = domain.createEnvironment(envName)
+  }
+  variable.setValue(value, environment)
+  return variable
+}
+
+methods.convertVariableIntoEnvironmentVariable = (domain, variable, key) => {
+  const envVariable = domain.createEnvironmentVariable(key)
+  const updateVariable = currify(methods.updateEnvironmentVariableWithEnvironmentValue, domain)
+  return variable.get('values').reduce(updateVariable, envVariable)
+}
+
+methods.addVariablesToVariableDomain = (context, domain, api) => {
+  const convertVariable = currify(methods.convertVariableIntoEnvironmentVariable, domain)
+  const vars = api.getIn([ 'store', 'variable' ]).map(convertVariable)
+
+  return new Store({
+    variable: vars
+  })
+}
+
+methods.createEnvironments = (context, api) => {
+  let store = new Store()
+  if (methods.needsStandardEnvironmentDomain(api)) {
+    const domain = methods.createStandardEnvironmentDomain(context, api)
+    store = methods.addVariablesToStandardDomain(context, domain, api)
+  }
+
+  if (methods.needsVariableEnvironmentDomain(api)) {
+    const domain = methods.createVariableEnvironmentDomain(context, api)
+    const variableStore = methods.addVariablesToVariableDomain(context, domain, api)
+    store = store.set('variable', variableStore.get('variable'))
+  }
+
+  return store
+}
+
+// TODO deal with case where there's an overlay for the url
+methods.convertEndpointsAndPathnameIntoDS = (pawRequest, store, endpoints, pathname) => {
+  const converted = endpoints.map((endpoint) => {
+    if (endpoint instanceof Reference) {
+      const variable = store.getIn([ 'endpoint', endpoint.get('uuid') ])
+      if (variable) {
+        return variable.createDynamicString()
+      }
+      return null
+    }
+
+    return methods.createEndpointDynamicString(endpoint)
+  }).filter(value => !!value)
+    .valueSeq()
+    .toJS()
+
+  if (converted.length === 1) {
+    return new DynamicString(converted[0], pathname)
+  }
+
+  const dv = pawRequest.addVariable('endpoint', converted[0], 'the endpoint of this url')
+  dv.schema = JSON.stringify({ type: 'string', enum: converted })
+
+  return new DynamicString(dv, pathname)
+}
+
+methods.convertRequestIntoPawRequest = (context, store, path, request) => {
+  const pathname = path.toURLObject(List([ '{', '}' ])).pathname
+  const name = request.get('name') || pathname
+  const method = request.get('method').toUpperCase()
+  const endpoints = request.get('endpoints')
+  const description = request.get('description') || ''
+
+  const pawRequest = context.createRequest(name, method, new DynamicString(), description)
+  const url = methods.convertEndpointsAndPathnameIntoDS(pawRequest, store, endpoints, pathname)
+  pawRequest.url = url
+  return pawRequest
+}
+
+// NOTE: not sure this is the best idea
+methods.convertResourceIntoGroup = (context, store, resource) => {
+  const path = resource.get('path')
+
+  const group = context.createRequestGroup(resource.get('name') || 'resource-group')
+
+  const convertRequest = currify(
+    methods.convertRequestIntoPawRequest,
+    context, store, path
+  )
+
+  return resource.get('methods')
+    .map(convertRequest)
+    .reduce(($group, pawRequest) => {
+      $group.appendChild(pawRequest)
+      return $group
+    }, group)
+}
+
+methods.createRequests = (context, store, api) => {
+  const convertResource = currify(methods.convertResourceIntoGroup, context, store)
+  const resources = api.get('resources').map(convertResource)
+
+  return resources
+}
+
+methods.serialize = ({ context, items, options } = {}, api) => {
+  const store = methods.createEnvironments(context, api)
+  const resources = methods.createRequests(context, store, api)
+  return true
+}
+
+export const __internals__ = methods
+export default PawSerializer
